BACKEND/requirements.txt
Flask==2.3.0
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
flask-jwt-extended==4.6.0
werkzeug==2.3.0
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
